requirements_production.txt
# 프로덕션 환경용 패키지
Django==5.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
psycopg[binary]==3.2.10
dj-database-url==3.0.1
python-dotenv==1.0.0

